+++ Editor/Core/Configuration/ProjectConfigurationBuilder.cs
@@ -5,7 +5,6 @@ using System.Linq;
 using UnityEditor;
 using UnityEditor.Build;
 using UnityEditor.Build.Reporting;
-using UnityEngine;
 
 namespace Unity.Services.Core.Configuration.Editor
 {