+++ Runtime/Configuration/AssemblyInfo.cs
@@ -5,11 +5,12 @@ using UnityEngine.Scripting;
 
 [assembly: InternalsVisibleTo("Unity.Services.Core.Registration")]
 [assembly: InternalsVisibleTo("Unity.Services.Core.Configuration.Editor")]
+[assembly: InternalsVisibleTo("Unity.Services.Core.TestUtils")]
 
 // Test assemblies
 #if UNITY_INCLUDE_TESTS
+[assembly: InternalsVisibleTo("Unity.Services.Core.EditorTests")]
 [assembly: InternalsVisibleTo("Unity.Services.Core.Tests")]
-[assembly: InternalsVisibleTo("Unity.Services.Core.TestUtils")]
 [assembly: InternalsVisibleTo("Unity.Services.Core.TestUtils.Tests")]
 [assembly: InternalsVisibleTo("Unity.Services.Core.TestUtils.EditorTests")]
 [assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")]